Week5 (21-25)/Day 24/ecommerce_semantic_search/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
pydantic==2.11.7
python-dotenv==1.1.1
numpy==2.2.6
faiss-cpu==1.12.0
sentence-transformers==5.1.0
openai==1.99.9
